Update manifest class to match discussed specification (#28)
This PR updates the manifest class to match the discussion on the
dataset doc. Let's focus all further discussion on the manifest in this
PR.

- I included a schema defining the manifest
- Data sources has been renamed to subsets
- The manifest now defines which fields / columns are available in a
subset and their types
- The location for each subset is resolved with the base path defined in
the metadata
- I'm no longer leveraging `dataclasses-json` as I don't just want to
build a 1 to 1 mapping of the json in Python
- All unused keys have been removed from the manifest

I'd like to see a similar PR for the component specification as well.