Add Fondant Dataset and Component class (#33)
This PR adds the `FondantDataset` wrapper class around the `Manifest`,
which loads data as Dask dataframes and allows to upload Dask dataframes
back to the cloud.

To test everything, the PR also includes a pipeline called "simple
pipeline" that includes 3 components: loading from hub, image filtering
and embedding. Each component needs to overwrite the `FondantComponent`
class, which uses the `FondantDataset` class behind the scenes.

To be discussed:

- [ ] I've added `project_name` to the metadata of the manifest, in
order to know the name of the cloud project. This is needed to
load/upload data using `fsspec`.

To do:

- [ ] for now I'm manually adding the `"gcs://"` prefix and `".parquet"`
suffix when loading data to and from the cloud, this needs to be
addressed (we need a cleaner way that is not hardcoded)
- [ ] only first 2 components are implemented, embedding component is to
do
- [ ] for the moment I'm still manually creating the KubeFlow component
yaml file for each component. This should be updated to automatically
create it based on the Fondant spec using the
[write_kubeflow_specification](https://github.com/ml6team/express/blob/db5807ae868fe36091d8d7f0061450312ab7477b/express/component_spec.py#L207)
method
- [ ] nicer way of creating and passing metadata. The only metadata that
is different per component is the `component_id`. Ideally getting rid of
`args.metadata`
- [ ] enforce usage of data types defined in output subsets when
creating the dataset (currently only the column names are checked)

---------

Co-authored-by: Niels Rogge <nielsrogge@Nielss-MacBook-Pro.local>
Co-authored-by: Philippe Moussalli <philippe.moussalli95@gmail.com>