Extract kubeflow logic from ComponentSpec to KubeflowComponent (#36)
This PR extracts all the kubeflow logic from the `ComponentSpec` to the
`KubeflowComponent` so the `ComponentSpec` is a minimal representation
of the fondant component specification. I also brought the classes a bit
more in line with the `Manifest` class.

This is preparation for the work to deduct the output manifest from the
input manifest and component spec only.