Evolve manifest based on component spec (#41)
Don't be alarmed by the size of this PR, it contains a lot of test cases
:sweat_smile:

This PR adds functionality to generate the output manifest based on the
input manifest and component specification alone. By deducting the
output manifest this way, it can be used for validation of the output
data in the component, and for static validation of the pipeline before
execution.

I implemented this as an `.evolve()` method on the manifest, since the
input manifest is always the starting point. This will also allow for
nice chaining in the future. The method is quite long, but I don't think
splitting it up will improve readability.

This PR only provides the functionality and tests. As a next step, we'll
have to integrate this into the `Dataset` class to actually leverage the
functionality for output validation.