Split fondant component into load and transform (#48)
PR that splits the Fondant Component into load and transform component.
The user will now call `FondantLoaderComponent` or
`FondantTransoromComponent` depending on the type of job that is run.

A few notes: 

- Both components rely on the `FondantComponentSpec` which requires both
`input_subsets` and defines an `input_manifest_path` as required
argument (based on `KubeflowComponentSpec`). In practice, the loader
component should not require both inputs. Wondering if it's worth
splitting the Spec to loader and transform as well and initialize both
`FondantLoaderComponent` and `FondantTransformComponent` with those.

- Currently only tested for the `FondantLoaderComponent`. I think the
`FondantTransformComponent` still needs to be implemented but part of
another PR (todos are still present in the code)