Split dataframe to subsets (#58)
This PR optimizes the process of executing dask transformations and
writing by creating delayed tasks that are executed in parallel. This is
mainly to optimize the process of writing many different subsets at the
same time.

I also added more documentation for the `Dataset` class. The changes did
require some refactoring