Handle kfp types (#108)
This PR enables the user to pass different `bool`, `lists` and `dict` to
a component. Kubeflow typically handles those arguments by serializing
them as a string . For this reason, they need to be de-serialized again
within the component in order for them to be properly handled.

This might go away once we move to V2. 

References to the issue: 
https://github.com/kubeflow/pipelines/issues/7457
https://github.com/kubeflow/pipelines/issues/7719