[Alpha pipeline] Write to hub component (#140)
PR that adds the write to hub component. 

Right now we're running `to_parquet` method twice, once within the
component itself to write to the hub and another time in fondant to
write the dataset. We should work on implementing a `DataWriter` class
huggingface/huggingface_hub#138.

Another thing that we need to configure is how to properly pass secrets
huggingface/huggingface_hub#139. Currently the hf token is passed as an
argument that is visible in kubeflow to all users who have access to it.

Linking also the issue for properly representing submitted images on the
hub (right now they are represented as byte strings:
https://github.com/huggingface/datasets/issues/5869