Add embedding based LAION retrieval component (#152)
This branch is based on the image-embedding branch which has a lot of
changes. I would suggest to first merge that PR which will give a much
smaller Diff here.

This component implements the LAION image retrieval component which uses
CLIP embeddings from the input subset to query the LAION database. It
returns an images subset with URLs, similar to the other prompt based
Clip Retrieval component. These URLs should then be downloaded by the
already-made image-downloading component.

---------

Co-authored-by: Philippe Moussalli <philippe.moussalli95@gmail.com>