Make LAION retrieval concurrent and use LAION ids (#168)
This PR tackles 2 changes:

1. It makes the LAION retrieval components concurrent, which speeds them
up significantly. To really use this at scale, we should set up our own
backend though.

As part of this change, I also included the LAION ids instead of a range
index, which leads to the second change:

2. Update the embedding, caption, and segmentation components to track
the ids correctly. Previously a new range index was created at the end,
which only coincidentally matched the input index. To track the ids, I
refactored these components to leverage the dataframe instead of delayed
API from Dask.

All changes are completely contained in there component, so the
components can be reviewed individually.