Custom component spec (#191)
PR that enables adding a custom component spec for each component as
described in #174

The implemented approach differs slightly from the described issue since
we forgot to take into account that we need to parse the arguments
twice. Once to retrieve the component spec and another time for the
other arguments which include things like the `metadata` and `manifest`
as well as use specific arguments (which are defined as part of the
component spec)

* The component spec is now an exact match to the one passed to the
pipeline. We no longer are defining it during the build process. This
enables us to use dynamic specs.
* We still keep the method `from_file()` method as this is used when
building and compiling the pipeline.

We still need to adapt the currently available read and write components
to be generic but this should be tackled in a separate PR.