Enable default component arguments (#199)
PR that enables adding default arguments as discussed in #179. 

Users can now define default arguments in the component specs. Those
arguments do not have to be explicitly defined in the `ComponentOp` but
could still be overridden.

Note: please review this
[PR](https://github.com/ml6team/fondant/pull/196) beforehand since I am
branched off from it.

Created a separate ticket to do the necessary changes #198. Best to
handle it in a separate PR.

---------

Co-authored-by: Robbe Sneyders <robbe.sneyders@gmail.com>