Add Pandas interface (#200)
Fixes #183 

There's some todo's left, we should
- [ ] Look into the redefinition of the divisions after we clear them.
Now that we take this out of the hands of the user, we should define
which strategy we want to follow here. (#205)
- [ ] Move to `hierarchical columns`. Pandas can work with hierarchical
columns, which would be a lot nicer as a user interface. I want to check
if I can make this work with Dask, and otherwise move the translation
from underscored names to hierarchical columns and back at the level of
the `PandasTransformComponent` (#204)
- [ ] Update the reusable components to leverage the
`PandasTransformComponent` (#203)