Enable optional component arguments (#201)
PR that enables defining optional arguments to the component spec. This
is needed to enable building reusable components.

For example, the `load_from_hub` component will need to format images
from hf format to bytes and this can be achieved by passing a specific
argument (`image_column_name`). This can be achieved either by:

- Passing `None` explicitly as an argument value at the `ComponentOp`
level
- Assigning `None` as the `default` argument field in the component
spec.

Previous attempt was to achieve this with the [optional field of the
kubeflow
spec](https://www.kubeflow.org/docs/components/pipelines/v1/reference/component-spec/#:~:text=Kubeflow%20Pipelines%20SDK.-,optional,-%3A%20Specifies%20if%20input)
but this returned many errors and also requires adding in an additional
`optional` field to the fondant spec. The current approach achieves the
same outcome with a simpler approach.

The `component_spec` argument is now also defined as an optional
argument. This allows use to define a component using the `from_file()`
method without having to explicitly define that argument. We still
return a valid error if `from_args()` is used and no `component_spec`
argument is provided since it defaults to `None`
[link](https://github.com/ml6team/fondant/blob/e90ac782a8c253720456fcbd05a9ced086ca5547/fondant/component.py#L61)