Provide hierarchical columns in pandas component (#211)
Fixes #204 

Unfortunately, wasn't able to propagate this through Dask. I got the
Dask code to work with hierarchical columns as long as the index was not
hierarchical, but I run into issues when trying to write this data to
parquet.

So this means that only the pandas component gets hierarchical columns
and I do the translation when we switch from Dask to Pandas.

I would propose to change our `_` for the flat column names to a symbol
which we expect to be less frequently used in the names of data columns
or fields. Eg. `images_data` could become `images+data` instead. This
will have no impact on users using the `PandasTransformComponent`, but
it will change for users using the `DaskTransformComponent` (which is
not clear yet how much there would be).

Retrieving the data using hierarchical columns is easy, but building a
dataframe with hierarchical columns is harder, we might want to add some
"cookbook" style examples for this.