Migrate stable diffusion & controlnet components to PandasTransformComponent (#219)
Contributes to #203 

This PR updates the reusable components of the stable diffusion /
controlnet pipeline to use the new Pandas interface. I submit these
separately since I've already tested them.

Some simplifications are clear:
- No more `map_partitions` needed for batching
- No more `meta` needed for `apply` or similar methods

Since we now set the `meta` of the `map_partitions` in the
`PandasTransformComponent` based on the component spec, the returned
dataframes are validated quite strictly. It does not allow any
additional columns.

The Pandas implementation can probably be improved further as I'm not
the biggest Pandas pro.