Build dev images on main (#236)
Fixes #220 

This PR adds a github actions workflow that builds our reusable images
on every merge to main. This should mitigate some of the issues we're
having with outdated fondant versions in the components.

Each image is tagged with the associated commit and we move the `dev`
tag to always point to the latest images build from main. This is now
the default version included in the `fondant_component.yaml` of the
reusable components. Just note that your runner might not recognize that
the image associated with the `dev` tag has changed, and you might have
to force it to pull the latest version.

The workflow uses [registry
caching](https://docs.docker.com/build/cache/backends/registry/) to
cache the docker images, which brings down the build time from
[~40m](https://github.com/ml6team/fondant/actions/runs/5398796173) to
[~20m](https://github.com/ml6team/fondant/actions/runs/5399309438). I
opted for registry caching as local github action caching is limited to
10GB, which would be too limited for the amount and size of images we're
building.

We can look into some additional improvements to this:
- [ ] **Direct the runners to always pull images**
This will introduce a small delay, but might be worth it to prevent hard
to detect issues because of outdated images.
- [ ] **Build fondant from local path**
The workflow introduced in this PR still requires changes to be merged
to main before they are built into the images. Especially while
developing on fondant itself, this might still be lacking in usability,
as we'd like to be able to test our local changes. This is possible by
passing an additional
[`--build-context`](https://docs.docker.com/engine/reference/commandline/buildx_build/#build-context),
but we'd have to integrate this into the local runner to really be
useful, without impacting the user experience.
- [ ] **Cache pip**
We can also leverage pip caching on top of / instead of docker caching.
This wouldn't add a lot of benefit on top of the current approach, as
the time spent on downloading packages and building wheels is negligible
compared to the time spent communicating with the registry cache. It
could be useful in allowing us to include the fondant dependency in the
`requirements.txt` again, which would be a bit more transparent. Then
the "Install dependencies" layer would no longer be cached by docker,
since we change the fondant version on each build, but the other
dependencies would be cached using pip. I believe this would still
increase the build time though as the cached packages still need to be
installed. To leverage the pip cache across containers, we can use a
[cache
mount](https://docs.docker.com/build/guide/mounts/#add-a-cache-mount).