Promote package from test.PyPI to PyPI without rebuilding (#258)
Fixes #255 

We were rebuilding the fondant package when publishing to PyPI, which
doesn't follow the "build once" mantra, and we were already suffering
the consequences. Our `build_components.sh` script updates the image
version in the component spec of the reusable components, which are
included in the fondant python package. However this script only runs in
the pipeline pushing to test.PyPI, not in the actual release pipeline
pushing to PyPI.

This PR now downloads the package from test.PyPI and pushes it to PyPI
instead of rebuilding, which should fix this issue.

Since the pipelines diverged further, I split them into separate files
and removed the conditionals.