Always pull images in local runner (#279)
Fixes #259 

This PR:
- Instructs the local runner to always pull images. This makes sure that
moving tags like `dev` or `latest` are up to date before running the
pipeline. As far as I can tell from my testing, this doesn't cause a
noticeable delay when the image was not updated.
- Adds the pipeline name to the docker compose yaml and removes orphan
containers. Containers are namespaced by the pipeline name, which
defaults to the cwd if not specified. So I set the name explicitly so we
don't accidentally remove containers from another pipeline.