Bugfix: data explorer artifact mounting (#310)
During the last fix I have incorporate a new bug in mounting the correct
path to the docker container.
I have adapted the implementation. The path in `fondant explore
--data-directory <path>` can also be an relative path now.
The path will be resolved to an absolute path. Inside the docker
container the folder will be located in the root folder.

---------

Co-authored-by: NielsRogge <48327001+NielsRogge@users.noreply.github.com>
Co-authored-by: Robbe Sneyders <robbe.sneyders@gmail.com>