Add int64 dtype (#338)
When running the Datacomp pipeline at scale, I encountered an outlier
image which had a width of about 37,000 pixels. As the component was
using int16 as the dtype of the width column, this failed as it was
outside the range. The error suggested to use int64 instead.

Then I noticed int64 is not supported yet, hence this PR adds it.