Bugfix data-explorer images (#382)
PR that fixes the issue with rendering images in the explorer. Images
currently have type `binary` and not `object`. This was changed to
render them properly.

Caveat: not all `binary` types will correspond to images so we will need
to implement another logic for image detection but this fix should do
for now.

Fixed another small bug related to errors in rendering the numerical
explorer when no numeric columns are present