Handle different base paths explorer (#427)
Linked to #426 #424 

Previous implementation of the local runner could only work with a local
runner and not the remote one. This PR enables the explorer to handle
both. If the base path is local, it will be mounted to the explorer.
Otherwise, the base path will be passed as a container argument inside
the component

Notes: 
* Cloud credentials are optional for the local runner (manifest is local
but can point to a remote ) but mandatory with the remote base path