Incorporate dask client (#410)
PR addressing #395.
Incorporate the usage of the dask client. By default, a LocalCluster will be initialized and used.
The LocalCluster provides a dashboard that enables enhanced debugging. To enable access to the
dashboard, we expose the default dashboard port in the docker-compose file, allowing it to be used
with the local runner.
---------

Co-authored-by: Philippe Moussalli <philippe.moussalli95@gmail.com>