Fix v2 defaults (#436)
PR that fixes the v2 default /optional types. KFP v2 has two different
optional types:

- Optional values that default to None -> defined by setting
`isOptional` to True and not passing the default value
- Optional values that default to another values -> defined by setting
the value in `defaultValue` and having `isOptional` False or empty.

This PR adjusts the componentSpec accordingly. Another nice win is that
we do not have to define bools/dicts/lists as strings in the component
spec

Few things no note: 

* Due to strict type. Mixing different data types is not allowed -> we
had to change the type of `input_partition_rows` to int. Which means it
does not accept the value `disable` anymore. Need to fix this in a
seperate PR
* 'inf` as float is not allowed