Set dask local scheduler as default (#438)
There might be some issues with the usage of the dask client and
local cluster. To make it more intuitive, we should revert to using the
old way as the default. This PR sets the old way as the default option,
and the client usage will now need to be set explicitly.