Add caching dependency (#479)
Current implemented caching does not take into account the previous
component cache key when estimating the cache key. This can be an issue
in some scenarios, for example:

 **1st run:** load_from_hub (input_dataset=x) -> download_images
 **2nd run:** load_from_hub (input_dataset=y) -> download_images
 
In the second run, the download images component won't execute and will
be cached since it might have the same input arguments as the first run.
However, both runs start from different input datasets and thus the
second run would actually end up loading in the dataset cached from the
1st run.

This PR addresses this by including the hash key of the dependent
component in the component cache key estimation.