Improve hit rate of download images component (#470)
This PR contains two changes to improve the hit rate of the download
images component:

- Introduce a Semaphore to limit the amount of concurrent requests. This
should help with the timeout errors. If this slows down the speed too
much, we can increase the number. If the changes in this PR work, we
probably want to make this an argument to the component.
- Configure httpx to follow redirects, [which it doesn't do by
default](https://www.python-httpx.org/compatibility/#redirects), but
requests / urllib3 does.