Move container registry to DockerHub (#514)
Fixes #495 

This PR moves everything to use DockerHub as our container registry
instead of Github container registry. This is necessary to support
managed pipeline services such as Vertex AI and Sagemaker.

As an improvement, the README of the component is now pushed as well.
[Example](https://hub.docker.com/r/fndnt/download_images).

The "fondant" organization is unfortunately taken on DockerHub, and we
can only request a transfer if we hold a US trademark on the name. Maybe
something for the future. For now, I went with "fndnt".