Add componentOp warning for unused configuration (#551)
Addresses #518 

Separating the ComponentOp for different runners is not ideal since it
would require changing the class names when switching runners. Instead,
it's best to return warning logs if a certain configuration is set but
not valid for a certain runner (e.g. nodepool for docker runner). This
PR modifies the `set_configuration` method of the Compiler to handle
this.