Add document viewer to dataset explorer (#666)
PR that adds document viewer to data explorer. The viewer attempts to
automatically detect the string type and render it to its most
appropriate format. Note that the normal string value does not get
rendered to a pdf value but rather a normal text/markdown.

I don't think it makes a lot of sense to render normal unformated text
to pdf. We can still detect whether a strings is a base64 pdf encoded
string and render that to it's original pdf format.

![Screenshot from 2023-11-22
17-32-32](https://github.com/ml6team/fondant/assets/47530815/2aa3a163-71dc-4dd2-b1d7-06f5d2d0450c)