Hide partitions from users (#677)
PR that hides the partitions from the users. Users can now scroll
through the dataset through a next/previous buttons.

In addition, a small changes was made to the statistics pages.
Previously it had a mix of both local and global statistics which made
it confusing. Now it's only restricted to global statistics.


https://github.com/ml6team/fondant/assets/47530815/05a32c54-7fbe-45c3-b7ba-d13e976674e2