Build to AWS ECR on release (#698)
The [prep-release pipeline
failed](https://github.com/ml6team/fondant/actions/runs/7086335306)
since it uses the `build_component.sh` script but doesn't login to AWS.

This PR fixes that.

I did not update the release pipeline. It uses the `tag_components.sh`
script, which adds a `latest` tag to the version being released. We
would need to update the script to enable updating the ECR tag as well.
But
- I don't think we're actually using the `latest` tag anywhere
- The `latest` tag would raise additional issues with the pull-through
cache, since it's cached
So I propose to leave it as is for now.