Support pipeline factory functions as CLI reference (#699)
In our RAG use case, [we want to start using a pipeline factory
function](https://github.com/ml6team/fondant-usecase-RAG/pull/34#discussion_r1414114415),
but this is currently not supported by the CLI.

This PR fixes that.

I also included a change (the first commit) which removes the `finally`
blocks. We were using this as follows:
```python
try:
    ref = ...
except SpecificError:
    ref = ...
finally:
    run(ref)
```

But apparently the `finally` block is also entered if any
non-`SpecificError` exception is raised in the `try` block. This lead to
all non-handled errors to become an `UnboundLocal` error. This is now no
longer the case, and the original error is not caught.