Unify operation and component spec (#741)
Fixes #728 

We now construct the operation spec once during the compilation process,
serialize it and pass it to the component as an argument. No need to
pass in the `produces` and `consumes` arguments anymore.

Some attributes like the arguments and the previous index that are part
of the component spec can still be accessed through the operation spec
since the component spec is it's attribute.