Add component argument inference (#763)
Fixes #751 

This PR introduces functionality to infer the arguments from a
`Component` class. The result is a dictionary with the argument names as
keys, and `Argument` instances as values, which is the format of
[`component_spec.args`.](https://github.com/ml6team/fondant/blob/8e828441eec8ff91074e5c8ccf16fe405b719594/src/fondant/core/component_spec.py#L193)

We can leverage this behavior for Lightweight Python components as
described in #750.

Did some TDD here, let me know if I missed any cases.