Fix ragas component (#759)
Fix to issue that @mrchtr raised

* `model_name` was actually referring to the class name used to import
the model which can be `OpenAI` or `ChatOpenAI`. This caused some issues
because we were trying to set `model_name` to `gpt-4`. Changed the
arguments and default to reflect that.

* ChatOpenAI seems to be the new default of defining models, changed the
default model class to that one

When initializing GPT-4 with `OpenAI()`:

```
UserWarning: You are trying to use a chat model. This way of initializing it is no longer supported. Instead, please use: `from langchain_community.chat_models import ChatOpenAI`
```

Also how it's set in RAGAS
[LINK](https://docs.ragas.io/en/latest/howtos/customisations/llms.html#:~:text=Copy%20code-,from%20langchain.chat_models%20import%20ChatOpenAI,-gpt4%20%3D%20ChatOpenAI)

Changes will be required to RAG pipeline after new release