Update caching arguments to include `ComponentOp` and the `Image` class (#802)
PR that updates the cache key estimation to take into account the
`ComponentSpec` instead of the `ComponentOp` (operation op has both the
component spec and the inner `produces` and `consumes` as well as the
`Image`

Tested on the example pipeline and caching seems to work fine for
lightweight component, only issue is that it will always download the
requirements (fondant + extra user dependencies) before skipping the
execution of the component. This can become frustrating for iterative
development with an increase of dependencies and components.

We might be able to mitigate this by providing base images with
pre-installed Fondant version and move the installation of the extras
after the cache hit estimation. Best to tackle this in a separate PR.
Issue ticket [here](https://github.com/ml6team/fondant/issues/805)

The PR also include some small changes to function names to best reflect
what they do